Change file structure, add skeletons.
